requirements.txt

opencv-python-headless
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.1.3
cmake==3.22.4
cvzone==1.5.6
Django==4.0.4
dlib==19.18.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.1.0
idna==3.3
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.1.0